+++ rmw_tickle/rmw_tickle/src/rmw_service.c
@@ -135,6 +135,9 @@ rmw_service_t* rmw_create_service(const rmw_node_t* node, const rosidl_service_t
         RMW_SET_ERROR_MSG("Expected implementation identifier to be " RMW_TICKLE_IDENTIFIER);
         return NULL;
     }
+    if (rmw_tickle_validate_qos_profile(qos_policies, false) != RMW_RET_OK) {
+        return NULL; // error message already set
+    }
 
     rmw_tickle_service_typesupport_t callbacks;
     if (!rmw_tickle_get_service_callbacks(type_support, &callbacks)) {